Deleted csp directives module as the abstraction was needless and OTT. Functionality moved into core csp module.
