+++ src/csp/mod.rs
@@ -3,13 +3,13 @@
 //
 // ToDo: Uses Threads, this may be overkill.
 mod directive;
-mod directives;
 mod domains;
 
 use crate::directives::GetDirectives;
 use crate::domains::Collection;
 use crate::parse;
 use serde_json::error;
+use std::thread;
 use std::thread::JoinHandle;
 
 // Collect the generated directives and compile them into a CSP string.
@@ -18,7 +18,7 @@ fn directives_to_csp(directives: Vec<JoinHandle<String>>) -> String {
 
     for directive in directives {
         let directive_string = directive.join().unwrap();
-        
+
         if !directive_string.is_empty() {
             csp.push_str(directive_string.as_str());
             csp.push_str(" ");
@@ -28,14 +28,31 @@ fn directives_to_csp(directives: Vec<JoinHandle<String>>) -> String {
     csp.trim().to_string()
 }
 
+// Make each directive based on the domains collection supplied and the
+// directives config.
+fn make_directives(directives_config: Vec<String>, domains: Collection) -> Vec<JoinHandle<String>> {
+    let mut directives: Vec<JoinHandle<String>> = vec![];
+
+    for directive_item in directives_config {
+        let domains_clone = domains.clone();
+
+        directives.push(
+            thread::spawn(move || directive::generate(directive_item, domains_clone))
+        );
+    }
+
+    directives
+}
+
 // Parse the JSON config and generate the Content Security Policy.
 pub fn generate(directives_config: impl GetDirectives, json: &str) -> Result<String, error::Error> {
     let domains: Result<Collection, error::Error> = parse::json(json);
 
     match domains {
         Ok(domains) => {
-            let directives: Vec<JoinHandle<String>> =
-                directives::generate(directives_config.get_directives(), domains);
+            let config = directives_config.get_directives();
+
+            let directives: Vec<JoinHandle<String>> = make_directives(config, domains);
 
             Ok(directives_to_csp(directives))
         }
